This is the first commit
Basic functionality is there.  Current support for PDF, PNG, EPS, SVG and Tif is implemented, but a bug in QT 4.4.0 is causing problems for PNG and Tif.  A lot of code cleaning is required and some changes to the way rotations are handled.  Labels are implemented, but are flaky and need some drastic improvements.  The current parser handles only xyz files right now.