Changed atom drawing defaults
The atom size is divided by a larger number, to allow more fine tuning.  Consequently, the default has been bumped up to 14 from 12.  The simple colored rendering has been redefined to be a mirror image of the HoukMol style.  I don't like the way the arcs run in HoukMol - the label often clashes with it.  The atom label is now centered and the paint routine has been cleaned a little to make it a little bit more readable.