Added DrawingDisplay class, changed resize policy, implemented background color functions
This turned out to be a much bigger change than anticipated.  The DrawingDisplay class, which is derived from QGraphicsView, was introduced to allow resize events to be handled.  This should make it easier to implement a crop function.  The background color and opacity can now be set, and it seems to work correctly.  I moved some code around in the atom class paint function earlier, but I just realized that the label doesn't show up in colored mode, so more work to be done there.  The default sizes are all screwed up at the moment, so I have to fix them some time...